scripts/requirements.txt
rembg==2.0.59
Pillow>=10.0.0
